move static versioning to separate file
