+++ Source/Versions.swift
@@ -8,81 +8,6 @@
 
 import Foundation
 
-extension Array {
-
-  func at(index: Int) -> T? {
-    if index >= 0 && index < self.count {
-      return self[index]
-    }
-    return nil
-  }
-}
-
-public enum Patch {
-  case NumberVersion(Int)
-  case StringVersion(String)
-
-  public init?(_ patch: String?) {
-    switch patch {
-      case let .Some(x): switch x {
-        case let num where num.toInt() != nil: self = .NumberVersion(num.toInt()!)
-        default: self = .StringVersion(x)
-      }
-      case .None: return nil
-    }
-  }
-
-  var number: Int? {
-    switch self {
-      case let .NumberVersion(x): return x
-      default: return nil
-    }
-  }
-
-  var string: String? {
-    switch self {
-      case let .StringVersion(s): return s
-      default: return nil
-    }
-  }
-}
-
-
-public struct Version : Equatable,  Comparable{
-
-  let major: Int?
-  let minor: Int?
-  let patch: Patch?
-  let string: String?
-
-  public init(_ version: String) {
-    let parts: Array<String> = split(version) { $0 == "." }
-    major = parts.at(0)?.toInt()
-    minor = parts.at(1)?.toInt()
-    patch = Patch(parts.at(2))
-    string = version
-  }
-}
-
-//MARK: - Equatable
-
-public func == (lhs: Version, rhs: Version) -> Bool {
-  return lhs.string == rhs.string
-}
-
-public func == (lhs: Version, rhs: Version?) -> Bool {
-  switch (rhs) {
-    case let .Some(r): return lhs.string == r.string
-    case .None: return false
-  }
-}
-
-//MARK: - Comparable
-
-public func < (lhs: Version, rhs: Version) -> Bool {
-  return lhs.string!.compare(rhs.string!, options: NSStringCompareOptions.NumericSearch) == .OrderedAscending
-}
-
 //MARK: - Versions string functionality
 
 public enum Semantic {
