+++ Versions Tests/Versions_Tests.swift
@@ -65,22 +65,4 @@ class Versions_Tests: XCTestCase {
         XCTAssertTrue("1.1.0".patchChange("1.1.1"))
         XCTAssertFalse("1.1.1".patchChange("1.1.1"))
     }
-
-  func testVersionStruct() {
-    let v = Version("1.2.3")
-
-    XCTAssertEqual(v.major!, 1)
-    XCTAssertEqual(v.minor!, 2)
-    XCTAssertEqual(v.patch!.number!, 3)
-    XCTAssertNil(v.patch?.string)
-  }
-
-  func testCompare() {
-    let v = Version("1.2.3")
-    let v2 = Version("1.2.4")
-    let vSame = Version("1.2.3")
-
-    XCTAssertFalse(v == v2)
-    XCTAssertTrue(v == vSame)
-  }
 }
